Move outer_puzzles.py
